Changed intersect function to use float math instead of double math

+++ Source/RTEngine/Shape.hpp
@@ -27,7 +27,7 @@ namespace RealRT
 			Shape(const Material &mat);
 			virtual ~Shape(void) = default;
 																			
-			virtual double Intersect(const Ray &incident, bool &flipNormals) const = 0; //override to return the distance from the origin to the point of intersection
+			virtual float Intersect(const Ray &incident, bool &flipNormals) const = 0; //override to return the distance from the origin to the point of intersection
 			virtual Vector3D Normal(const Vector3D &pt) const = 0;	//override to get the UNIT normal at the given point
 
 			virtual bool IsLight(void) const;  //return true if this shape is a light source
